DNS: moved messages implicit conversions from actors package to messages package
